Test disabling of strong params
With strong params enabled and a seemingly valid payload, errors are
returned when updating a contact: "Param not allowed" for name_first,
name_last, and email.

This error is resolved by setting the jsonapi-resources.rb initializer
config.raise_if_parameters_not_allowed to false and reproducable when
set to true.